Refactor controller and services with exception, add placeInReview endpoints

+++ src/main/kotlin/hu/zsof/restaurantApp/service/PlaceService.kt
@@ -3,11 +3,13 @@ package hu.zsof.restaurantApp.service
 import hu.zsof.restaurantApp.dto.FilterDto
 import hu.zsof.restaurantApp.dto.PlaceDto
 import hu.zsof.restaurantApp.dto.PlaceInReviewDto
+import hu.zsof.restaurantApp.exception.MyException
 import hu.zsof.restaurantApp.model.Place
 import hu.zsof.restaurantApp.model.PlaceInReview
 import hu.zsof.restaurantApp.model.convertToDto
 import hu.zsof.restaurantApp.repository.PlaceInReviewRepository
 import hu.zsof.restaurantApp.repository.PlaceRepository
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.util.*
@@ -15,7 +17,7 @@ import java.util.*
 @Service
 @Transactional
 class PlaceService(private val placeRepository: PlaceRepository, private val placeInReviewRepository: PlaceInReviewRepository) {
-    fun newPlace(newPlace: Place): PlaceDto {
+    fun newPlace(newPlace: Place): Place {
         val theNewPlace = Place()
         theNewPlace.name = newPlace.name
         theNewPlace.address = newPlace.address
@@ -32,43 +34,43 @@ class PlaceService(private val placeRepository: PlaceRepository, private val pla
         theNewPlace.rate = 0.0f
         theNewPlace.usersNumber = 0
         theNewPlace.user = newPlace.user
-        return placeRepository.save(theNewPlace).convertToDto()
+        return placeRepository.save(theNewPlace)
     }
 
     fun getAllPlace(): MutableList<Place> = placeRepository.findAll()
 
-    fun getPlaceById(id: Long) = placeRepository.findById(id)
-
-    fun deletePlaceByIdByUser(id: Long, creatorId: Long): Boolean {
-        val place = placeRepository.findById(id).get()
-
-        //TODO jó így?? törölni csak a sajátot lehet
-        if (place.user.id == creatorId) {
-
-            return if (placeRepository.existsById(id)) {
-                placeRepository.deleteById(id)
-                true
-            } else {
-                false
-            }
-        }
-        return false
-    }
-
-    fun deletePlaceById(id: Long): Boolean {
-        return if (placeRepository.existsById(id)) {
-            placeRepository.deleteById(id)
-            true
+    fun getPlaceById(id: Long): Place {
+        val place = placeRepository.findById(id)
+        if (place.isPresent) {
+            return place.get()
         } else {
-            false
+            throw MyException("Place not found", HttpStatus.NOT_FOUND)
         }
-
     }
 
-    fun deleteAll() = placeRepository.deleteAll()
+    fun deletePlaceByIdByUser(placeId: Long, creatorId: Long) {
+        val place = getPlaceById(placeId)
+
+        if (place.user.id == creatorId) {
+            deletePlaceById(placeId)
+        } else {
+            throw MyException("User has no permission to delete this place", HttpStatus.FORBIDDEN)
+        }
+    }
+
+
+    fun deletePlaceById(id: Long) {
+        if (placeRepository.existsById(id)) {
+            placeRepository.deleteById(id)
+        } else {
+            throw MyException("Place not found", HttpStatus.NOT_FOUND)
+        }
+    }
+
     fun filterPlaces(filterItems: FilterDto): MutableList<Place> {
         val getAllFilteredPlace = getAllPlace().filter { restaurantList ->
             filterItems.filter.convertToList().compare(restaurantList.filter.convertToList())
+            // todo ide nem lehet ? -> filterItems.type== restaurantList.type
         }.toMutableList()
 
         return getAllFilteredPlace.filter { restaurantList ->
@@ -77,33 +79,40 @@ class PlaceService(private val placeRepository: PlaceRepository, private val pla
     }
 
 
-    fun updatePlace(place: Place): Optional<PlaceInReviewDto> {
+    fun updatePlace(place: Place, creatorId: Long): PlaceInReviewDto {
+        // todo modosítás oldal ahol jová lehet hagyni modosításokat - eredeti és módosított dolgok egymás mellett látszanak, nem kerül ki rendes oldalra a módosítás amig nem fogadta admin el --> el kell addig tárolni
+        if (place.user.id == creatorId) {
 
-        // Add place to PlaceInReview table with changes
-        val newPlaceInReviewWithUpdates = PlaceInReview()
-        newPlaceInReviewWithUpdates.name = place.name
-        newPlaceInReviewWithUpdates.address = place.address
-        newPlaceInReviewWithUpdates.type = place.type
-        newPlaceInReviewWithUpdates.price = place.price
-        newPlaceInReviewWithUpdates.image = place.image
-        newPlaceInReviewWithUpdates.filter = place.filter
-        newPlaceInReviewWithUpdates.phoneNumber = place.phoneNumber
-        newPlaceInReviewWithUpdates.email = place.email
-        newPlaceInReviewWithUpdates.web = place.web
-        newPlaceInReviewWithUpdates.latitude = place.latitude
-        newPlaceInReviewWithUpdates.longitude = place.longitude
-        newPlaceInReviewWithUpdates.openDetails = place.openDetails
-        newPlaceInReviewWithUpdates.rate = 0.0f
-        newPlaceInReviewWithUpdates.usersNumber = 0
-        newPlaceInReviewWithUpdates.user = place.user
+            // Add place to PlaceInReview table with changes
+            val newPlaceInReviewWithUpdates = PlaceInReview()
+            newPlaceInReviewWithUpdates.name = place.name
+            newPlaceInReviewWithUpdates.address = place.address
+            newPlaceInReviewWithUpdates.type = place.type
+            newPlaceInReviewWithUpdates.price = place.price
+            newPlaceInReviewWithUpdates.image = place.image
+            newPlaceInReviewWithUpdates.filter = place.filter
+            newPlaceInReviewWithUpdates.phoneNumber = place.phoneNumber
+            newPlaceInReviewWithUpdates.email = place.email
+            newPlaceInReviewWithUpdates.web = place.web
+            newPlaceInReviewWithUpdates.latitude = place.latitude
+            newPlaceInReviewWithUpdates.longitude = place.longitude
+            newPlaceInReviewWithUpdates.openDetails = place.openDetails
+            newPlaceInReviewWithUpdates.rate = 0.0f
+            newPlaceInReviewWithUpdates.usersNumber = 0
+            newPlaceInReviewWithUpdates.user = place.user
 
-        //Save place to PlaceInReview table
-        val placeInReviewWithUpdates = placeInReviewRepository.save(newPlaceInReviewWithUpdates).convertToDto()
+            //Save place to PlaceInReview table
+            val placeInReviewWithUpdates = placeInReviewRepository.save(newPlaceInReviewWithUpdates).convertToDto()
 
-        //Delete place from Place table
-        placeRepository.deleteById(place.id)
+            //Delete place from Place table
+            // todo id generált, megegyik a placein review idja a place rendes idjával? valószínuleg nem
+            placeRepository.deleteById(place.id)
 
-        // Add modified place to PlaceInReview able
-        return Optional.of(placeInReviewWithUpdates)
+            // Add modified place to PlaceInReview able
+            return placeInReviewWithUpdates
+        } else {
+            throw MyException("User has no permission to update this place", HttpStatus.FORBIDDEN)
+        }
+
     }
 }
